debugging for blockpc test case
commented : copy map part
Fixed: pc2.ll and alive-bug.ll

Problem in alive-bug: The brnach terminator inst in BB's Pred can have any operator besides ICMP.
We need to just do default thing for anything other than ICMP in BR inst. The reason to do this
is because isKnownNonEqual() can be called for ICMP inst, not for AND, OR operators and so on.

Problem in pc2.ll -> some BB got skipped when the candidates were checked. Fixed that.

New prob found in logic of stamping out the BB's. We currently look for Pred's stamp to make a decision
of proceeding to addPCs for current BB. But in case of BlockPCs, we will eventually be left with all BB's
stamped out! Thus, we wont be able to find any optimization. We should look for :
1. Look for cur BB stamp in addPC() when you reach case EQ or case NE
2. In EQ case, if result = false; then u mark: Cur = true and pred->successor[1] = false !! instead of marking pred=false
   If res = true; mark: cur = false; pred->successor[1] = true
3. likewise do for NE case
4. initialization of all BB's stamp was done true in ExtractCands() function earlier. Now, look for if BB->stamp is false
   skip the BB. Otherwise, mark the stamp = true and proceed further to call addPC.