improve path conditions by using data flow function isKnownNonEqual
This patch improves the basic block traversal for the cases where branching
occurs. Usually for branch condition, we have to take both True and False branch,
and return optimizations as found by Souper.
Now, by using the data flow function: isKnownNonEqual(), we can make a decision
to take either of the True or False branches and thus improving the basic block
traversal.
The data flow function can be used only for EQ or NE operators involved
in branch conditions. This idea so far helps in making blockpc's more strict.