add else case in guessing logic
non-negative guessing logic tests if sign bit is known to be 0, followed
by guessing if sign bit is 1 and if it fails in both cases, it used to set
NonNegative to 0 which would make souper infer the results as non-negative.
We now added a third case, if guessing fails for both 0 and 1 sign bit value,
it would still set nonNegative as 1. so that souper can never infer the data flow
fact as non-negative (n).